refactor: modularize frontend testing utilities
