+++ src/web/frontend/frontend_testing.py
@@ -1,35 +1,17 @@
 #!/usr/bin/env python3
+"""Frontend testing infrastructure.
+
+Provides a unified entry point for running component, routing,
+state management and integration tests for the frontend system.
 """
-Frontend Testing Infrastructure
-Agent_Cellphone_V2_Repository TDD Integration Project
+from __future__ import annotations
 
-This module provides comprehensive testing infrastructure for the frontend system:
-- Component testing utilities
-- Routing testing utilities
-- State management testing
-- Integration testing helpers
-- Mock data generators
-- Test fixtures and utilities
-
-Author: Web Development & UI Framework Specialist
-License: MIT
-"""
-
-import json
 import logging
-import pytest
-import asyncio
-
-from src.utils.stability_improvements import stability_manager, safe_import
-from pathlib import Path
-from typing import Dict, List, Any, Optional, Callable, Union
-from dataclasses import dataclass, asdict
 from datetime import datetime
-from unittest.mock import Mock, MagicMock, patch, AsyncMock
-import tempfile
-import shutil
+from typing import Dict, List, Optional
 
-# Frontend imports
+import pytest
+
 from .frontend_app import (
     FlaskFrontendApp,
     FastAPIFrontendApp,
@@ -45,51 +27,25 @@ from .frontend_router import (
     NavigationState,
     RouteGuard,
     RouteMiddleware,
-    create_router_with_default_routes,
     RouteBuilder,
+    create_router_with_default_routes,
 )
+from .component_testing import run_all_component_tests, run_single_component_test
+from .routing_testing import run_routing_tests as perform_routing_tests
+from .state_testing import run_state_management_tests
+from .integration_testing import run_integration_tests as perform_integration_tests
+from .testing_models import TestResult, TestSuite
+from .testing_utils import MockDataGenerator, TestUtilities
 
-# Configure logging
+# expose fixtures from testing_utils
+pytest_plugins = ["src.web.frontend.testing_utils"]
+
 logging.basicConfig(level=logging.INFO)
 logger = logging.getLogger(__name__)
 
-# ============================================================================
-# TESTING MODELS & DATA STRUCTURES
-# ============================================================================
-
-
-@dataclass
-class TestResult:
-    """Result of a frontend test"""
-
-    test_name: str
-    test_type: str
-    status: str  # 'passed', 'failed', 'skipped'
-    duration: float
-    error_message: Optional[str]
-    component_tested: Optional[str]
-    route_tested: Optional[str]
-    timestamp: datetime
-    metadata: Dict[str, Any]
-
-
-@dataclass
-class TestSuite:
-    """Collection of related tests"""
-
-    name: str
-    description: str
-    tests: List[TestResult]
-    total_tests: int
-    passed_tests: int
-    failed_tests: int
-    skipped_tests: int
-    total_duration: float
-    created_at: datetime
-
 
 class FrontendTestRunner:
-    """Main frontend test runner"""
+    """Main frontend test runner."""
 
     def __init__(self):
         self.test_results: List[TestResult] = []
@@ -97,11 +53,9 @@ class FrontendTestRunner:
         self.current_suite: Optional[TestSuite] = None
         self.mock_data_generator = MockDataGenerator()
         self.test_utilities = TestUtilities()
-
         logger.info("Frontend test runner initialized")
 
     def run_component_tests(self, component_name: str = None) -> TestSuite:
-        """Run component tests"""
         suite_name = f"component_tests_{datetime.now().strftime('%Y%m%d_%H%M%S')}"
         suite = TestSuite(
             name=suite_name,
@@ -114,33 +68,23 @@ class FrontendTestRunner:
             total_duration=0.0,
             created_at=datetime.now(),
         )
-
         self.current_suite = suite
-
         try:
-            # Run component tests
             if component_name:
-                self._run_single_component_test(component_name)
+                run_single_component_test(component_name, suite)
             else:
-                self._run_all_component_tests()
-
-            # Calculate suite statistics
+                run_all_component_tests(suite)
             self._calculate_suite_stats(suite)
-
-            # Store suite
             self.test_suites[suite_name] = suite
-
             logger.info(
                 f"Component tests completed: {suite.passed_tests}/{suite.total_tests} passed"
             )
             return suite
-
-        except Exception as e:
+        except Exception as e:  # pragma: no cover - safety
             logger.error(f"Error running component tests: {e}")
             raise
 
     def run_routing_tests(self) -> TestSuite:
-        """Run routing tests"""
         suite_name = f"routing_tests_{datetime.now().strftime('%Y%m%d_%H%M%S')}"
         suite = TestSuite(
             name=suite_name,
@@ -153,35 +97,24 @@ class FrontendTestRunner:
             total_duration=0.0,
             created_at=datetime.now(),
         )
-
         self.current_suite = suite
-
         try:
-            # Run routing tests
-            self._run_routing_tests()
-
-            # Calculate suite statistics
+            perform_routing_tests(suite)
             self._calculate_suite_stats(suite)
-
-            # Store suite
             self.test_suites[suite_name] = suite
-
             logger.info(
                 f"Routing tests completed: {suite.passed_tests}/{suite.total_tests} passed"
             )
             return suite
-
-        except Exception as e:
+        except Exception as e:  # pragma: no cover - safety
             logger.error(f"Error running routing tests: {e}")
             raise
 
-    def run_integration_tests(self) -> TestSuite:
-        """Run integration tests"""
-        suite_name = f"integration_tests_{datetime.now().strftime('%Y%m%d_%H%M%S')}"
+    def run_state_tests(self) -> TestSuite:
+        suite_name = f"state_tests_{datetime.now().strftime('%Y%m%d_%H%M%S')}"
         suite = TestSuite(
             name=suite_name,
-            description="Integration testing suite",
-            tests=[],
+            description="State management testing suite",
             tests=[],
             total_tests=0,
             passed_tests=0,
@@ -190,219 +123,66 @@ class FrontendTestRunner:
             total_duration=0.0,
             created_at=datetime.now(),
         )
-
         self.current_suite = suite
-
         try:
-            # Run integration tests
-            self._run_integration_tests()
-
-            # Calculate suite statistics
+            run_state_management_tests(suite)
             self._calculate_suite_stats(suite)
-
-            # Store suite
             self.test_suites[suite_name] = suite
+            logger.info(
+                f"State tests completed: {suite.passed_tests}/{suite.total_tests} passed"
+            )
+            return suite
+        except Exception as e:  # pragma: no cover - safety
+            logger.error(f"Error running state tests: {e}")
+            raise
 
+    def run_integration_tests(self) -> TestSuite:
+        suite_name = f"integration_tests_{datetime.now().strftime('%Y%m%d_%H%M%S')}"
+        suite = TestSuite(
+            name=suite_name,
+            description="Integration testing suite",
+            tests=[],
+            total_tests=0,
+            passed_tests=0,
+            failed_tests=0,
+            skipped_tests=0,
+            total_duration=0.0,
+            created_at=datetime.now(),
+        )
+        self.current_suite = suite
+        try:
+            perform_integration_tests(suite)
+            self._calculate_suite_stats(suite)
+            self.test_suites[suite_name] = suite
             logger.info(
                 f"Integration tests completed: {suite.passed_tests}/{suite.total_tests} passed"
             )
             return suite
-
-        except Exception as e:
+        except Exception as e:  # pragma: no cover - safety
             logger.error(f"Error running integration tests: {e}")
             raise
 
     def run_all_tests(self) -> Dict[str, TestSuite]:
-        """Run all test suites"""
         try:
             logger.info("Starting comprehensive frontend testing...")
-
-            # Run all test suites
             component_suite = self.run_component_tests()
             routing_suite = self.run_routing_tests()
+            state_suite = self.run_state_tests()
             integration_suite = self.run_integration_tests()
-
             all_suites = {
                 "component": component_suite,
                 "routing": routing_suite,
+                "state": state_suite,
                 "integration": integration_suite,
             }
-
-            # Generate summary report
             self._generate_summary_report(all_suites)
-
             logger.info("All frontend tests completed")
             return all_suites
-
-        except Exception as e:
+        except Exception as e:  # pragma: no cover - safety
             logger.error(f"Error running all tests: {e}")
             raise
 
-    def _run_single_component_test(self, component_name: str):
-        """Run tests for a single component"""
-        start_time = datetime.now()
-
-        try:
-            # Test component creation
-            component = create_component(component_name, {"test": True})
-            assert component.type == component_name
-            assert component.props["test"] is True
-
-            # Test component state management
-            component.state["test_state"] = "test_value"
-            assert component.state["test_state"] == "test_value"
-
-            # Record test result
-            duration = (datetime.now() - start_time).total_seconds()
-            test_result = TestResult(
-                test_name=f"test_{component_name}_creation",
-                test_type="component",
-                status="passed",
-                duration=duration,
-                error_message=None,
-                component_tested=component_name,
-                route_tested=None,
-                timestamp=datetime.now(),
-                metadata={"test_type": "creation"},
-            )
-
-            self.current_suite.tests.append(test_result)
-
-        except Exception as e:
-            duration = (datetime.now() - start_time).total_seconds()
-            test_result = TestResult(
-                test_name=f"test_{component_name}_creation",
-                test_type="component",
-                status="failed",
-                duration=duration,
-                error_message=str(e),
-                component_tested=component_name,
-                route_tested=None,
-                timestamp=datetime.now(),
-                metadata={"test_type": "creation"},
-            )
-
-            self.current_suite.tests.append(test_result)
-
-    def _run_all_component_tests(self):
-        """Run tests for all components"""
-        # Test basic components
-        basic_components = ["Button", "Card", "Input", "Modal"]
-
-        for component_name in basic_components:
-            self._run_single_component_test(component_name)
-
-    def _run_routing_tests(self):
-        """Run routing tests"""
-        start_time = datetime.now()
-
-        try:
-            # Test router creation
-            router = create_router_with_default_routes()
-            assert router is not None
-            assert len(router.routes) > 0
-
-            # Test route matching
-            home_match = router.match_url("/")
-            assert home_match is not None
-            assert home_match["matched"] is True
-
-            # Test navigation
-            success = router.navigate_to("/dashboard")
-            assert success is True
-
-            # Test breadcrumbs
-            breadcrumbs = router.get_breadcrumbs()
-            assert len(breadcrumbs) > 0
-
-            # Record test result
-            duration = (datetime.now() - start_time).total_seconds()
-            test_result = TestResult(
-                test_name="test_routing_basic",
-                test_type="routing",
-                status="passed",
-                duration=duration,
-                error_message=None,
-                component_tested=None,
-                route_tested="/",
-                timestamp=datetime.now(),
-                metadata={"test_type": "basic_routing"},
-            )
-
-            self.current_suite.tests.append(test_result)
-
-        except Exception as e:
-            duration = (datetime.now() - start_time).total_seconds()
-            test_result = TestResult(
-                test_name="test_routing_basic",
-                test_type="routing",
-                status="failed",
-                duration=duration,
-                error_message=str(e),
-                component_tested=None,
-                route_tested="/",
-                timestamp=datetime.now(),
-                metadata={"test_type": "basic_routing"},
-            )
-
-            self.current_suite.tests.append(test_result)
-
-    def _run_integration_tests(self):
-        """Run integration tests"""
-        start_time = datetime.now()
-
-        try:
-            # Test Flask frontend app
-            flask_app = FrontendAppFactory.create_flask_app()
-            assert flask_app is not None
-            assert hasattr(flask_app, "app")
-            assert hasattr(flask_app, "component_registry")
-
-            # Test FastAPI frontend app
-            fastapi_app = FrontendAppFactory.create_fastapi_app()
-            assert fastapi_app is not None
-            assert hasattr(fastapi_app, "app")
-            assert hasattr(fastapi_app, "component_registry")
-
-            # Test component registry integration
-            registry = flask_app.component_registry
-            registry.register_component("TestComponent", lambda x: x)
-            assert "TestComponent" in registry.list_components()
-
-            # Record test result
-            duration = (datetime.now() - start_time).total_seconds()
-            test_result = TestResult(
-                test_name="test_integration_apps",
-                test_type="integration",
-                status="passed",
-                duration=duration,
-                error_message=None,
-                component_tested=None,
-                route_tested=None,
-                timestamp=datetime.now(),
-                metadata={"test_type": "app_integration"},
-            )
-
-            self.current_suite.tests.append(test_result)
-
-        except Exception as e:
-            duration = (datetime.now() - start_time).total_seconds()
-            test_result = TestResult(
-                test_name="test_integration_apps",
-                test_type="integration",
-                status="failed",
-                duration=duration,
-                error_message=str(e),
-                component_tested=None,
-                route_tested=None,
-                timestamp=datetime.now(),
-                metadata={"test_type": "app_integration"},
-            )
-
-            self.current_suite.tests.append(test_result)
-
     def _calculate_suite_stats(self, suite: TestSuite):
-        """Calculate statistics for a test suite"""
         suite.total_tests = len(suite.tests)
         suite.passed_tests = len([t for t in suite.tests if t.status == "passed"])
         suite.failed_tests = len([t for t in suite.tests if t.status == "failed"])
@@ -410,7 +190,6 @@ class FrontendTestRunner:
         suite.total_duration = sum(t.duration for t in suite.tests)
 
     def _generate_summary_report(self, suites: Dict[str, TestSuite]):
-        """Generate a summary report of all test suites"""
         total_tests = sum(s.total_tests for s in suites.values())
         total_passed = sum(s.passed_tests for s in suites.values())
         total_failed = sum(s.failed_tests for s in suites.values())
@@ -426,227 +205,20 @@ class FrontendTestRunner:
         logger.info(f"Skipped: {total_skipped}")
         logger.info(f"Total Duration: {total_duration:.2f}s")
         logger.info("=" * 60)
-
         for suite_name, suite in suites.items():
             logger.info(
                 f"{suite_name.upper()}: {suite.passed_tests}/{suite.total_tests} passed"
             )
-
         logger.info("=" * 60)
 
 
-# ============================================================================
-# TEST UTILITIES
-# ============================================================================
-
-
-class TestUtilities:
-    """Utility functions for testing"""
-
-    def create_mock_component(
-        self, component_type: str = "TestComponent"
-    ) -> UIComponent:
-        """Create a mock component for testing"""
-        return create_component(
-            component_type,
-            {
-                "id": "test-id",
-                "className": "test-class",
-                "data-testid": "test-component",
-            },
-        )
-
-    def create_mock_route(self, path: str = "/test") -> RouteConfig:
-        """Create a mock route for testing"""
-        return RouteConfig(
-            path=path,
-            name="test-route",
-            component="TestComponent",
-            props={"title": "Test Page"},
-            meta={"requiresAuth": False},
-            children=[],
-            guards=[],
-            middleware=[],
-            lazy_load=False,
-            cache=True,
-            created_at=datetime.now(),
-            updated_at=datetime.now(),
-        )
-
-    def create_mock_navigation_state(self) -> NavigationState:
-        """Create a mock navigation state for testing"""
-        return NavigationState(
-            current_route="/test",
-            previous_route="/",
-            route_params={"id": "123"},
-            query_params={"page": "1"},
-            navigation_history=["/", "/test"],
-            breadcrumbs=[
-                {"path": "/", "name": "Home"},
-                {"path": "/test", "name": "Test"},
-            ],
-            timestamp=datetime.now(),
-        )
-
-    def assert_component_props(
-        self, component: UIComponent, expected_props: Dict[str, Any]
-    ):
-        """Assert component has expected properties"""
-        for key, value in expected_props.items():
-            assert (
-                component.props[key] == value
-            ), f"Expected {key}={value}, got {component.props.get(key)}"
-
-    def assert_route_config(
-        self, route: RouteConfig, expected_path: str, expected_component: str
-    ):
-        """Assert route has expected configuration"""
-        assert (
-            route.path == expected_path
-        ), f"Expected path {expected_path}, got {route.path}"
-        assert (
-            route.component == expected_component
-        ), f"Expected component {expected_component}, got {route.component}"
-
-    def assert_navigation_state(self, state: NavigationState, expected_route: str):
-        """Assert navigation state has expected route"""
-        assert (
-            state.current_route == expected_route
-        ), f"Expected route {expected_route}, got {state.current_route}"
-
-
-class MockDataGenerator:
-    """Generates mock data for testing"""
-
-    def generate_mock_user(self) -> Dict[str, Any]:
-        """Generate mock user data"""
-        return {
-            "id": "user-123",
-            "username": "testuser",
-            "email": "test@example.com",
-            "role": "user",
-            "created_at": datetime.now().isoformat(),
-            "last_login": datetime.now().isoformat(),
-        }
-
-    def generate_mock_component_data(self, component_type: str) -> Dict[str, Any]:
-        """Generate mock component data"""
-        base_data = {
-            "id": f"{component_type.lower()}-123",
-            "className": f"{component_type.lower()}-component",
-            "data-testid": f"test-{component_type.lower()}",
-        }
-
-        if component_type == "Button":
-            base_data.update(
-                {"text": "Test Button", "type": "button", "disabled": False}
-            )
-        elif component_type == "Card":
-            base_data.update(
-                {
-                    "title": "Test Card",
-                    "content": "This is a test card component",
-                    "footer": "Card Footer",
-                }
-            )
-
-        return base_data
-
-    def generate_mock_route_data(self) -> List[Dict[str, Any]]:
-        """Generate mock route data"""
-        return [
-            {
-                "path": "/",
-                "name": "home",
-                "component": "HomePage",
-                "props": {"title": "Home"},
-                "meta": {"requiresAuth": False},
-            },
-            {
-                "path": "/dashboard",
-                "name": "dashboard",
-                "component": "DashboardPage",
-                "props": {"title": "Dashboard"},
-                "meta": {"requiresAuth": True},
-            },
-            {
-                "path": "/settings",
-                "name": "settings",
-                "component": "SettingsPage",
-                "props": {"title": "Settings"},
-                "meta": {"requiresAuth": True},
-            },
-        ]
-
-
-# ============================================================================
-# PYTEST FIXTURES
-# ============================================================================
-
-
-@pytest.fixture
-def frontend_test_runner():
-    """Fixture for frontend test runner"""
-    return FrontendTestRunner()
-
-
-@pytest.fixture
-def test_utilities():
-    """Fixture for test utilities"""
-    return TestUtilities()
-
-
-@pytest.fixture
-def mock_data_generator():
-    """Fixture for mock data generator"""
-    return MockDataGenerator()
-
-
-@pytest.fixture
-def flask_frontend_app():
-    """Fixture for Flask frontend app"""
-    return FrontendAppFactory.create_flask_app()
-
-
-@pytest.fixture
-def fastapi_frontend_app():
-    """Fixture for FastAPI frontend app"""
-    return FrontendAppFactory.create_fastapi_app()
-
-
-@pytest.fixture
-def frontend_router():
-    """Fixture for frontend router"""
-    return create_router_with_default_routes()
-
-
-@pytest.fixture
-def mock_component():
-    """Fixture for mock component"""
-    return TestUtilities().create_mock_component()
-
-
-@pytest.fixture
-def mock_route():
-    """Fixture for mock route"""
-    return TestUtilities().create_mock_route()
-
-
-@pytest.fixture
-def mock_navigation_state():
-    """Fixture for mock navigation state"""
-    return TestUtilities().create_mock_navigation_state()
-
-
-# ============================================================================
-# PYTEST TEST FUNCTIONS
-# ============================================================================
+# ---------------------------------------------------------------------------
+# Pytest test functions
+# ---------------------------------------------------------------------------
 
 
 def test_component_creation(test_utilities):
-    """Test component creation"""
     component = test_utilities.create_mock_component("TestButton")
-
     assert component.type == "TestButton"
     assert component.id is not None
     assert component.props["data-testid"] == "test-component"
@@ -655,19 +227,12 @@ def test_component_creation(test_utilities):
 
 
 def test_route_configuration(test_utilities):
-    """Test route configuration"""
     route = test_utilities.create_mock_route("/test-page")
-
     test_utilities.assert_route_config(route, "/test-page", "TestComponent")
-    assert route.name == "test-route"
-    assert route.props["title"] == "Test Page"
-    assert route.meta["requiresAuth"] is False
 
 
 def test_navigation_state(test_utilities):
-    """Test navigation state"""
     state = test_utilities.create_mock_navigation_state()
-
     test_utilities.assert_navigation_state(state, "/test")
     assert state.previous_route == "/"
     assert state.route_params["id"] == "123"
@@ -675,142 +240,71 @@ def test_navigation_state(test_utilities):
 
 
 def test_flask_frontend_app(flask_frontend_app):
-    """Test Flask frontend app creation"""
     assert flask_frontend_app is not None
     assert hasattr(flask_frontend_app, "app")
     assert hasattr(flask_frontend_app, "component_registry")
     assert hasattr(flask_frontend_app, "state_manager")
-
-    # Test component registry
     registry = flask_frontend_app.component_registry
     assert "Button" in registry.list_components()
     assert "Card" in registry.list_components()
 
 
 def test_fastapi_frontend_app(fastapi_frontend_app):
-    """Test FastAPI frontend app creation"""
     assert fastapi_frontend_app is not None
     assert hasattr(fastapi_frontend_app, "app")
     assert hasattr(fastapi_frontend_app, "component_registry")
     assert hasattr(fastapi_frontend_app, "state_manager")
-
-    # Test component registry
     registry = fastapi_frontend_app.component_registry
     assert "Button" in registry.list_components()
     assert "Card" in registry.list_components()
 
 
 def test_frontend_router(frontend_router):
-    """Test frontend router functionality"""
     assert frontend_router is not None
     assert len(frontend_router.routes) > 0
-
-    # Test route matching
     home_match = frontend_router.match_url("/")
     assert home_match is not None
     assert home_match["matched"] is True
-
-    # Test navigation
     success = frontend_router.navigate_to("/dashboard")
     assert success is True
-
-    # Test breadcrumbs
     breadcrumbs = frontend_router.get_breadcrumbs()
     assert len(breadcrumbs) > 0
 
 
 def test_component_registry_integration(flask_frontend_app):
-    """Test component registry integration"""
     registry = flask_frontend_app.component_registry
-
-    # Test component registration
     registry.register_component(
         "CustomComponent", lambda x: x, "template", "style", "script"
     )
     assert "CustomComponent" in registry.list_components()
-
-    # Test component retrieval
     component_func = registry.get_component("CustomComponent")
     assert component_func is not None
 
 
 def test_state_manager_integration(flask_frontend_app):
-    """Test state manager integration"""
     state_manager = flask_frontend_app.state_manager
-
-    # Test initial state
     initial_state = state_manager.get_state()
     assert initial_state.app_name == "Agent_Cellphone_V2 Frontend"
     assert initial_state.theme == "light"
-
-    # Test state update
     state_manager.update_state({"theme": "dark"})
     updated_state = state_manager.get_state()
     assert updated_state.theme == "dark"
 
 
-# ============================================================================
-# ASYNC TEST FUNCTIONS
-# ============================================================================
-
-
 @pytest.mark.asyncio
 async def test_fastapi_websocket_async(fastapi_frontend_app):
-    """Test FastAPI WebSocket functionality (async)"""
-    # This would typically test actual WebSocket connections
-    # For now, we'll test the app structure
     assert fastapi_frontend_app is not None
     assert hasattr(fastapi_frontend_app, "app")
 
 
-# ============================================================================
-# INTEGRATION TEST FUNCTIONS
-# ============================================================================
-
-
 def test_full_frontend_integration(flask_frontend_app, frontend_router):
-    """Test full frontend integration"""
-    # Test that all components work together
     app = flask_frontend_app
     router = frontend_router
-
-    # Test component registry
     registry = app.component_registry
     assert len(registry.list_components()) > 0
-
-    # Test routing
     assert len(router.routes) > 0
-
-    # Test state management
     state_manager = app.state_manager
     initial_state = state_manager.get_state()
     assert initial_state is not None
-
-    # Test navigation
     success = router.navigate_to("/dashboard")
     assert success is True
-
-
-# ============================================================================
-# MAIN EXECUTION
-# ============================================================================
-
-if __name__ == "__main__":
-    # Run tests directly if script is executed
-    print("Running frontend tests...")
-
-    test_runner = FrontendTestRunner()
-
-    try:
-        # Run all test suites
-        results = test_runner.run_all_tests()
-
-        print("\nTest execution completed successfully!")
-        print(f"Total test suites: {len(results)}")
-
-        for suite_name, suite in results.items():
-            print(f"{suite_name}: {suite.passed_tests}/{suite.total_tests} passed")
-
-    except Exception as e:
-        print(f"Test execution failed: {e}")
-        raise
